Define directory layout
Define the directory layout for the application, based on the layout of
the `flask` application. Also add the MIT license to the project. We
define `sdep` as containing the package, `tests` containing the tests,
`docs` as containing the docs, and `examples` as containing
documentation for deploying an example static website.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>