Add travis, Docker for dev, and dev Makefile
We will use `.travis.yml` for continuous testing. We made the decision
to run our tests on `travis` outside of a Docker container, so thus in
our Makefile we have `local_test` and `test`, where `local_test` is the
actual test command and `test` wraps the command in a Docker container.
We create a `Dockerfile` for development tasks for `sdep` as well as
outline the development requirements in `requirements.txt`. I think we
specify the actual package requirements for `sdep` elsewhere.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>