Add `click` as an `sdep` dependency
We will use the `click` library for creating the `sdep` CLI interface.
This commit also provides a template for how to specify and install
`sdep` dependencies in `setup.py`. This is opposed to development
dependencies which we specify in `requirements.txt` and are needed to
develop `sdep`, but not actually run it.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>