Remove Route53 and Cloudfront mentions
Initially, I thought that in addition to automating the configuration
of S3, I would also automate DNS with Route53 and caching with
Cloudfront. However, neither one of those platforms is particularly well
supported by boto or moto. In addition, manual configuration is simple
through Cloudflare, and also allows the end user more flexibility. So at
least for the near future, the decision was made to limit the scope of
`sdep` to `S3`, while providing documentation for integrating with
`Cloudflare `.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>