Reset semver
I want to version this project with `semver`. However, I was not
properly specifying the version number in that format, as I did not
specify a patch number. This commit basically starts the versioning over
at `0.1.0`, the first working release. I deleted all previous versions
on pypi and Github tags, meaning we are completely starting over at
`0.1.0`.

After this commit, we will more closely adhere to semver, as well as
utilizing the Changelog.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>